Correct the author
I understand that the editing the original paragraph written by me was necessary to make space for the other parts. However, the paragraph with the editing does not make sense to me any longer. I do not want to be associated with it. Here is the original:

In this section, we first briefly summarize the primary prior work, especially the analysis procedures, to clarify the differences between the original study and our own objectives and corresponding methods discussed subsequently. To address the scarcity of single-cell transcriptome data on the pancreas that associate gene expressions with cell types, Segerstolpe et al. [7] apply single-cell RNA sequencing to the tissue samples from the two cohorts of individuals with or without T2D and map the results to the latent space for cellar phenotyping. Their major contributions based on the analysis at the granularity of individual genes of identified cell types include: reporting correlations by Spearman's rho between genetic expressions and obesity in the group of the males without T2D; and finding significantly different gene expressions, comparing the two cohorts by ANOVA.